Create input state macro directly from Qiskit and remove custom meas
Moves to Qiskit circuits for state preparation within the classical shadow experiment.

This change replaces the custom macro-based approach with Qiskit's QuantumCircuit for defining input states, enabling better flexibility and integration with existing quantum algorithms and transpilation tools.
The possibility of specifying the measurement basis is now removed as it is currently not used in the main QUA program.